+++ Sophos/sophos_module/asset_connector/__init__.py
@@ -1,4 +1,3 @@
 from sophos_module.asset_connector.device_assets import SophosDeviceAssetConnector
 
 __all__ = ["SophosDeviceAssetConnector"]
-